refactor: extract archival list_archived into diesel_common
